Solution for 2023 1-1 and 1-2
Rust's regex crate doesn't support getting a list of overlapping
matches, so instead I used a second backwards regex and matched
each line in reverse order to get the final "digit". My original
submission for part 2 failed, and my theory was that the cause was
an overlapping "digit" such as "threeight" or "twone". Rust's regex
engine will return ["three"] instead of ["three", "eight"] in this case.
I haven't verified that this is actually the case, but after handling it
the code did return the correct answer.

Total solve time: ~75m