Solution for 2023 3-1 and 3-2
For this problem, I just walked the input by indexing it as a 2D array.
Tons of hideous bounds checking logic, this was not a very rusty
solution. Most of the logic that I used to walk around the array in part
1 could not be re-used for part 2.

At least this solution is fast, both parts finish in less
than 1ms each.

Total solve time: 80m (54m + 26m)