Solution for 2023 4-1 and 4-2
This one was relatively easy, but I spent a lot of time fighting with
Rust integer types (i64 vs u32 vs usize). I implemented a Card struct
that holds the card number, the set of winning numbers, and the set of
numbers present on the card. The card number itself turned out to not be
used in part 2, so could have been omitted. I was also able to split
this into a bunch of testable unit functions. My solution to part 2
isn't quite as nice as for part 1, but  proves that the answer to most
programming puzzles is just "use a HashMap".

Total solve time: 52m (31m + 21m)