Solution for 2023 8-1 and 8-2
Part 1 was simple to implement, just using a HashMap to move through the
nodes worked well. This could be optimized by using a HashMap of
randomly assigned numeric IDs in place of the strings (plus a marker for
initial/terminal node status), as my solution does a ton of string
copying.

Part 2's brute force solution was infeasible, so I ended up looking for
a hint, and learned that the number of steps for each path between
terminal nodes was the same for every loop; hence finding the lowest
common multiple of each path's loop length would suffice to solve this
problem.

Total solve time: 54m (33m + 21m)