Solution for 2023 11-1 and 11-2
For part 1, I took the approach of increasing the actual map size by
inserting additional rows and columns of blank space. This worked fine,
and matched the visuals shown in the puzzle page.

Part 2 is where this approach instantly became infeasible, as the empty
space now had to be made 1,000,000 times larger instead of just twice as
large. Even doing this with the test input would take too long and use
too much memory. Instead, I implemented the much simpler method of just
counting the number of empty rows and columns between each pair of
galaxies, and adding that to the taxicab distance calculated on the
unexpanded map. This is a scenario where the part 2 twist lead me to
make the part 1 solution much simpler as well.

Total solve time: 68m (54m + 14m)