Solution for 2023 13-1 and 13-2
This solution sort of came together as I wrote it, I wasn't entirely
sure if it would work when I started. As always, much of the time was
spent debugging offsets when doing array index math.

For part 2, I immediately knew that testing all potential smudge fixes
wouldn't be feasible, so I had the idea to allow for exactly one
inequality when comparing across potential mirror edges. If a match is
found without any "smudges", skip it; if an inequality has already been
found for a given mirror edge and a second is encountered, skip it.

Total solve time: 67m (59m + 8m)