Solution for 2023 14-1 (and probably 14-2, pending computation)
For part 1, I went the efficient route of not actually moving the
rolling rocks on the board. Unfortunately, that all went out the door
for part 2, where I need to simulate the movement of rocks after 1
billion full rotation cycles. I've got zero clue how to do this without
actually rolling rocks around the board a billion times.

I extrapolated from a checkpoint of 1 million cycles that this will
take about 10 hours to finish running, which is not as fast as 5-2's
execution time of 3.5m, but is still within the bounds of possibility
unlike 12-2's 670 billion years. So I'll leave it running overnight and
hopefully get the correct answer in the morning.

Total solve time: 66m (30m + 36m)