Solution for 2023 14-2
An hour or so later, I figured out the trick: the platform is bound to
fall into a loop well before 1 billion cycles. If I can track each
platform state along with the iteration count, and detect when the
platform's current state matches a past cycle, I can then figure out
which step of the loop will be encountered at iteration 1 billion.

This was simple to implement, and once I had it in place I was able to
see just how broken my old part 2 code was. This took a bit more time to
solve, because every single day I put myself into a mess of array index
math mistakes.

In the end, this now takes <20ms to execute instead of 10 hours, and as
a bonus will actually give the correct answer at the end!

Total solve time: 2h (30m + [36m + 54m])